requirements.txt
absl-py==1.2.0
albumentations==1.2.1
appnope==0.1.3
asttokens==2.0.5
async-generator==1.10
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2021.5.30
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
colorama==0.4.5
configparser==5.2.0
crayons==0.4.0
cryptography==37.0.4
cycler==0.10.0
decorator==5.1.1
executing==0.8.3
fonttools==4.34.4
future==0.18.2
glob2==0.7
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h11==0.13.0
idna==2.10
imageio==2.20.0
importlib-metadata==4.12.0
imutils==0.5.4
ipython==7.34.0
jedi==0.18.1
joblib==1.1.0
kiwisolver==1.3.1
lxml==4.9.1
Markdown==3.4.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
opencv-python-headless==4.2.0.32
outcome==1.2.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pefile==2022.5.30
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkg_resources==0.0.0
prompt-toolkit==3.0.30
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
roboflow==0.2.11
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2207130030
threadpoolctl==3.1.0
tifffile==2021.11.2
torch==1.8.0+cu111
torchaudio==0.8.0
torchtext==0.10.0
torchvision==0.9.0
tqdm==4.64.0
traitlets==5.3.0
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.6
validators==0.18.2
wcwidth==0.2.5
webdriver-manager==3.5.1
Werkzeug==2.1.2
wget==3.2
wsproto==1.1.0
zipp==3.8.1
